# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'lciii-scaffolding'","description":"'LCII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'LCII Initial Scaffolding'"}],"paths":{"/api/countries":{"get":{"tags":["country-controller"],"operationId":"getCountries","parameters":[{"name":"code","in":"query","required":false,"schema":{"type":"string"}},{"name":"name","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDto"}}}}}}},"post":{"tags":["country-controller"],"operationId":"saveCountries","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CountryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDto"}}}}}}}},"/api/{continent}/continent":{"get":{"tags":["country-controller"],"operationId":"getCountriesByContinent","parameters":[{"name":"continent","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDto"}}}}}}}},"/api/countries/{language}/language":{"get":{"tags":["country-controller"],"operationId":"getCountriesByLanguage","parameters":[{"name":"language","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDto"}}}}}}}},"/api/countries/most-borders":{"get":{"tags":["country-controller"],"operationId":"getCountriesWithMostBorders","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDto"}}}}}}}},"/api/countries/filtered":{"get":{"tags":["country-controller"],"operationId":"getCountriesFiltered","parameters":[{"name":"code","in":"query","required":false,"schema":{"type":"string"}},{"name":"name","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDto"}}}}}}}}},"components":{"schemas":{"CountryRequest":{"type":"object","properties":{"amountOfCountryToSave":{"type":"integer","format":"int32"}}},"CountryDto":{"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"}}}}}}